+++ source/artifacts/_artifactDictionary.js
@@ -17,6 +17,8 @@ const ROLL_TABLE = {
 for (const file of [
 	"amethystspyglass.js",
 	"bloodshieldsword.js",
+	"boatparts.js",
+	"celestialknightinsignia.js",
 	"crystalshard.js",
 	"enchantedmap.js",
 	"hammerspaceholster.js",