+++ source/artifacts/_artifact_blueprint.js
@@ -4,4 +4,4 @@ module.exports = new ArtifactTemplate("name",
 	"description",
 	"scaling description",
 	"element"
-).setFlavorText({ name: "heading", value: "text" })
+).setFlavorText({ name: "heading", value: "text" });